requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.5
pydantic-settings==2.7.1

# OpenAI
openai==1.10.0

# Web Search
tavily-python==0.3.0

# Vector Database (RAG)
numpy<2.0.0,>=1.22.5
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Generation
reportlab==4.0.9
Pillow==11.0.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0

# Date & Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
